chore: improvements by comments
