+++ components/Header/index.vue
@@ -1,26 +1,13 @@
 <script setup lang="ts">
 import Logo from "../UI/Logo.vue";
 import Navigation from "./Navigation.vue";
-import LoginModal from "./LoginModal.vue";
+import Profile from "@/components/Profile/Profile.vue";
 import { useScrollListener } from "./useScrollListener";
-import RegisterModal from "./RegisterModal.vue";
 
 const scrolled = ref(false);
-const isLoginModalOpen = ref(false);
-const isRegisterModalOpen = ref(false);
 
 if (import.meta.client) {
   useScrollListener(scrolled);
-}
-
-function openModal(modal: "login" | "register") {
-  if (modal === "login") {
-    isLoginModalOpen.value = true;
-    isRegisterModalOpen.value = false;
-  } else {
-    isLoginModalOpen.value = false;
-    isRegisterModalOpen.value = true;
-  }
 }
 </script>
 
@@ -33,17 +20,13 @@ function openModal(modal: "login" | "register") {
         : 'top-2 md:top-8'
     "
   >
-    <div class="container flex items-center">
-      <Logo class="mr-16" />
-      <Navigation />
-      <LoginModal
-        v-model:open="isLoginModalOpen"
-        @open-register-modal="openModal('register')"
-      />
-      <RegisterModal
-        v-model:open="isRegisterModalOpen"
-        @open-login-modal="openModal('login')"
-      />
+    <div class="container flex items-center justify-between">
+      <div class="flex items-center">
+        <Logo class="mr-16" />
+        <Navigation />
+      </div>
+
+      <Profile />
     </div>
   </header>
 </template>
