+++ components/Profile/ModalCloseButton.vue
@@ -1,3 +1,7 @@
+<script setup lang="ts">
+const open = defineModel<boolean>("open");
+</script>
+
 <template>
   <div class="absolute top-4 right-4 border-0">
     <UButton icon="i-heroicons-x-mark" variant="ghost" @click="open = false" />
